Refactor: change file name from task2 to fa_module
